Fix angular loading in the web-client
PR3381

This patch adds an alias in the webpack config for angular, which allows us to use the angular package that ships with angular-patternfly. Additionally, package.json was revised to remove dependencies that are included in angular-patternfly, or that aren't currently being used.

Reviewed-by: aazores
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2017-May/023107.html