Update jvm-list for web-gateway format
Notably, timestamps are not directly numbers, but objects with a key
"$numberLong" and the timestamp value as a string representing a Unix
timestamp. Additionally, the server response object is nested within a
"response" key. Finally, JVM class names are trimmed to exclude package names
in the jvm-list as a partial fix for main-classes overflowing the bounds of the
cards.

Reviewed-by: jkang
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2017-June/023419.html